Added BookingViewModel
