+++ HotelDemo/Core/View+buttonBack.swift
@@ -5,18 +5,18 @@
 //  Created by Лаборатория on 07.09.2023.
 //
 
-import SwiftUI
-
-extension View {
-
-    @ViewBuilder
-    func buttonBack(completion: @escaping ()->()) -> some View {
-        Button {
-            completion()
-        } label: {
-            Image(systemName: "chevron.backward")
-                .foregroundColor(.black)
-        }
-    }
-
-}
+//import SwiftUI
+//
+//extension View {
+//
+//    @ViewBuilder
+//    func buttonBack(completion: @escaping ()->()) -> some View {
+//        Button {
+//            completion()
+//        } label: {
+//            Image(systemName: "chevron.backward")
+//                .foregroundColor(.black)
+//        }
+//    }
+//
+//}
