+++ HotelDemo/ViewModel/BookingViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  BookingViewModel.swift
+//  HotelDemo
+//
+//  Created by Лаборатория on 08.09.2023.
+//
+
+import Foundation
+
+class BookingViewModel: ObservableObject {
+
+    let name: String = "Бронирование"
+    
+}
